1/10 - JPA-Database

+++ src/main/java/com/api_arquitetura_example/api_arquitetura_example/DataLoader.java
@@ -1,8 +1,10 @@
-import entity.Product;
+package com.api_arquitetura_example.api_arquitetura_example;
+
+import com.api_arquitetura_example.api_arquitetura_example.entity.Product;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
-import repository.ProductRepository;
+import com.api_arquitetura_example.api_arquitetura_example.repository.ProductRepository;
 
 import java.util.Arrays;
 import java.util.List;
@@ -27,14 +29,17 @@ public class DataLoader implements CommandLineRunner {
         System.out.println(">>> [READ] Total de produtos: " + products.size());
 
         // 3. UPDATE (Atualizar)
-        Product toUpdate = repository.findById(1L).get();
-        toUpdate.setName("Smartphone Ultra Pro");
-        repository.save(toUpdate);
-        System.out.println(">>> [UPDATE] Produto ID 1 atualizado!");
+        repository.findById(1L).ifPresent(toUpdate -> {
+            toUpdate.setName("Smartphone Ultra Pro");
+            repository.save(toUpdate);
+            System.out.println(">>> [UPDATE] Produto ID 1 atualizado!");
+        });
 
         // 4. DELETE (Deletar)
-        repository.deleteById(3L);
-        System.out.println(">>> [DELETE] Produto ID 3 removido!");
+        if (repository.existsById(3L)) {
+            repository.deleteById(3L);
+            System.out.println(">>> [DELETE] Produto ID 3 removido!");
+        }
 
         System.out.println(">>> Setup Inicial Concluído com sucesso!");
     }
